pyproject.toml
[tool.poetry]
name = "wikeyword"
version = "0.1.0"
description = ""
authors = ["kishimoto-banana <drehbleistift@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
wikipedia2vec = "^1.0.4"
flask-bootstrap = "^3.3.7"
mecab-python3 = "^0.996.5"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
